Makefile for cross compilation
Not completely as I want it yet. Needs some GNU Make wizardry to reduce
code dupe, and doesn't yet bundle GNU tools with Windows binaries
(mostly a problem of Where To Get Them).

Darwin builds are untested, Linux and Windows binaries are good so
committing anyway in the hope that it all just works.